+++ src/clj/matcher_combinators/clj_test.clj
@@ -212,7 +212,7 @@
         the-rest  (rest (rest form))
         roughly-delta?  `(fn [expected#]
                            (core/->PredMatcher (fn [actual#]
-                                                 (utils/roughly? expected# actual# ~delta))
+                                                 (utils/within-delta? ~delta expected# actual#))
                                                (str "roughly " expected# " (+/- " ~delta ")")))
         form' (concat [directive] the-rest)]
     `(if (not (= 3 (count '~(rest form))))
@@ -221,14 +221,6 @@
           :message  ~msg
           :expected (symbol (str "`" '~directive "` expects 3 arguments: a `delta` number, a `matcher`, and the `actual`"))
           :actual   (symbol (str (count '~(rest form)) " were provided: " '~form))})
-       ~(build-match-assert 'match-roughly?
-                            {java.lang.Integer    roughly-delta?
-                             java.lang.Short      roughly-delta?
-                             java.lang.Long       roughly-delta?
-                             java.lang.Float      roughly-delta?
-                             java.lang.Double     roughly-delta?
-                             java.math.BigDecimal roughly-delta?
-                             java.math.BigInteger roughly-delta?
-                             clojure.lang.BigInt  roughly-delta?}
+       ~(build-match-assert 'match-roughly? [number? roughly-delta?]
                             msg
                             form'))))